refactor: controller, service, repository 계층화 리팩토링
