+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,15 +1,9 @@
 package roomescape.controller;
 
 import java.net.URI;
-import java.sql.Date;
-import java.sql.PreparedStatement;
 import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -19,85 +13,37 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
 import roomescape.dto.ReservationRequest;
 import roomescape.dto.ReservationResponse;
+import roomescape.service.ReservationService;
 
 @RestController
 @RequestMapping("/reservations")
 public class ReservationController {
 
-    private final JdbcTemplate jdbcTemplate;
-    private final RowMapper<Reservation> actorRowMapper = (rs, rowNum) ->
-            new Reservation(
-                    rs.getLong("reservation_id"),
-                    rs.getString("name"),
-                    rs.getDate("date").toLocalDate(),
-                    new ReservationTime(rs.getLong("time_id"), rs.getTime("time_value").toLocalTime())
-            );
+    private final ReservationService reservationService;
 
-    public ReservationController(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
+    public ReservationController(final ReservationService reservationService) {
+        this.reservationService = reservationService;
     }
 
     @GetMapping
     @ResponseStatus(HttpStatus.OK)
     public List<ReservationResponse> findAll() {
-        String sql = """
-            SELECT 
-                r.id AS reservation_id,
-                r.name,
-                r.date,
-                t.id AS time_id,
-                t.start_at AS time_value
-            FROM reservation AS r
-            INNER JOIN reservation_time AS t
-            ON r.time_id = t.id""";  // time_value 이건 왜필요하지?
-        List<Reservation> reservations = jdbcTemplate.query(sql, actorRowMapper);
+        List<Reservation> reservations = reservationService.findAll();
         return ReservationResponse.fromReservations(reservations);
     }
 
-    private Reservation findById(final long id) {
-        String sql = """
-            SELECT 
-                r.id AS reservation_id,
-                r.name,
-                r.date,
-                t.id AS time_id,
-                t.start_at AS time_value
-            FROM reservation AS r
-            INNER JOIN reservation_time AS t
-            ON r.time_id = t.id
-            WHERE r.id = ?
-            """;
-        Reservation reservation = jdbcTemplate.queryForObject(sql, actorRowMapper, id);
-        return reservation;
-    }
-
     @PostMapping
-    public ResponseEntity<ReservationResponse> add(
-            @RequestBody final ReservationRequest reservationRequest) {
-        String sql = "INSERT INTO reservation (name, date, time_id) VALUES (?, ?, ?)";
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-
-        jdbcTemplate.update(connection -> {
-            PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
-            ps.setString(1, reservationRequest.name());
-            ps.setDate(2, Date.valueOf(reservationRequest.date()));
-            ps.setLong(3, reservationRequest.timeId());
-            return ps;
-        }, keyHolder);
-
-        long id = keyHolder.getKey().longValue();
-        Reservation reservation = findById(id);
+    public ResponseEntity<ReservationResponse> add(@RequestBody final ReservationRequest reservationRequest) {
+        Reservation reservation = reservationService.add(reservationRequest);
         ReservationResponse reservationResponse = ReservationResponse.from(reservation);
-        return ResponseEntity.created(URI.create("/reservations/" + id)).body(reservationResponse);
+        return ResponseEntity.created(URI.create("/reservations/" + reservation.getId())).body(reservationResponse);
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public int delete(@PathVariable(name = "id") final long id) {
-        String sql = "DELETE FROM reservation WHERE id = ?"; // TODO: 만약 찾을 수 없는 아이디면?
-        return jdbcTemplate.update(sql, id);
+        return reservationService.delete(id);
     }
 }
