+++ src/main/resources/templates/admin/time.html
@@ -22,10 +22,10 @@
   <div class="collapse navbar-collapse" id="navbarSupportedContent">
     <ul class="navbar-nav ml-auto">
       <li class="nav-item">
-        <a class="nav-link" href="/reservation">Reservation</a>
+        <a class="nav-link" href="/admin/reservation">Reservation</a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" href="/time">Time</a>
+        <a class="nav-link" href="/admin/time">Time</a>
       </li>
     </ul>
   </div>
